Move test using fresh_db to separate test module
